2009-08-21 ----------         added 'crusde -p install <plugin.so>' which is used in install_plugins.sh                 This is the new way to fill $CRUSDE_HOME/plugins/plugins.xml with the plugins that come                 with the standard distribution.
                install_plugins.sh is a tcsh script which contains lists of the standard plugins in the 
                order necessary to meet all dependencies. For each plugin
        
                        crusde -p install <plugin.so>

                is called which uses the the same functionality the existing GUI based PluginManager uses to test 
                the plugins and add them to the database, only that the GUI is not started. 

                - SimulationCore has installPlugin(string) added 
                - InputHandler has the respective checking added -- does not work with multiple options, yet!!

2009-07-11
----------
        added possibility to allow for optional parameters:
                - crusde-API expanded by crusde_register_optional_param* functions which ask for a default value for this parameter
                - ParamWrapper carries flag 'optional' which is set by the registration functions
                - InputHandler checks for this flag when going through the list of non-initialized (in XML file) params.
        added info.h
                - contains the existing usage() function, called in InputHandler.
                - contains new version() function, that gives ... tadaaa ... version information. (add retrieval of SVN revision number!)

        added output functions to API for plugins that are also used internally, all work like printf:
                - crusde_info(message, ...): writes message to stdout
                - crusde_warning(message, ...): writes warning message to stderr
                - crusde_error(message, ...): writes error message to stderr and terminates CrusDe
                - crusde_debug(message, ...): writes debug message to stderr if compiled with -DDEBUG

        added 'quiet' commandline option: -q/-Q
                - shuts info messages off, errors and warnings are still displayed, as well as debugging messages.



git-svn-id: svn://svn.gps.alaska.edu/crusde/trunk@6 1095105e-3e63-0410-b14e-63ce1eee999c

+++ src/crusde_api.cpp
@@ -31,6 +31,7 @@
 #include <iostream>
 #include <map>
 #include <string>
+#include <stdio.h>
 #include "crusde_api.h"
 #include "SimulationCore.h"
 #include "DataOutPlugin.h"
@@ -44,24 +45,102 @@ using namespace std;
 //void print_params(multimap<string, ParamWrapper> m);
 //void print_params_map(map<unsigned int, multimap<string, ParamWrapper> >m);
 
-
 /* PARAMETER REGISTRATION */
+//functions for parameter registration; all type specific functions do:
+//creates param wrapper, registers it with the simulation core and returns it to the
+//calling function so they can allocate memory for the right type.
 
+//error, warning and debugging functions
+extern "C" void crusde_error(const char* format, ...){
+	va_list args;
+	va_start( args, format );
+	fprintf( stderr, "[ Crusde ERROR ] : " );
+	vfprintf( stderr, format, args );
+	fprintf( stderr, "\n" );
+	va_end( args );
+	fflush( stderr );
+
+	//abort program
+	crusde_exit(-1);
+}
+
+extern "C" void crusde_warning(const char* format, ...){
+	va_list args;
+	va_start( args, format );
+	fprintf( stderr, "[ Crusde WARNING ] : " );
+	vfprintf( stderr, format, args );
+	fprintf( stderr, "\n" );
+	va_end( args );
+	fflush( stderr );
+}
+
+extern "C" void crusde_info(const char* format, ...){
+	if(!SimulationCore::instance()->isQuiet()){
+		va_list args;
+		va_start( args, format );
+		fprintf( stdout, "[ Crusde ] : " );
+		vfprintf( stdout, format, args );
+		fprintf( stdout, "\n" );
+		va_end( args );
+		fflush( stdout );
+	}
+}
+
+extern "C" void crusde_debug(const char* format, ...){
+#ifdef DEBUG
+	//cannot check for quietness, since one should be able to call it even in SimulationCore constructor.
+	va_list args;
+	va_start( args, format );
+	fprintf( stderr, "[ Crusde debug ] : " );
+	vfprintf( stderr, format, args );
+	fprintf( stderr, "\n" );
+	va_end( args );
+	fflush( stderr );
+#endif /* defined(DEBUG) */
+}
+
+
+//REGISTER DOUBLE
+//use this function if parameter is not optional
 extern "C" double* crusde_register_param_double(const char* name, PluginCategory category)
 {
-printf("crusde_register_param_double called for: %s\n", name);
+	crusde_debug("crusde_register_param_double called for: %s", name);
 	ParamWrapper *p = new ParamWrapper();
+	p->setOptional(false);
 	SimulationCore::instance()->registerParam(p, name, category);
 	double* add = p->newDouble();
 	return add;
 }
 
+//if you want parameter to be optional
+extern "C" double* crusde_register_optional_param_double(const char* name, PluginCategory category, double default_value)
+{
+crusde_debug("crusde_register_optional_param_double called for:%s", name);
+	ParamWrapper *p = new ParamWrapper();
+	p->setOptional(true);
+	SimulationCore::instance()->registerParam(p, name, category);
+	double* add = p->newDouble(default_value);
+	return add;
+}
+
+//REGISTER STRING
 extern "C" char** crusde_register_param_string(const char* name, PluginCategory category)
 {
-printf("crusde_register_param_string called for: %s\n", name);
+crusde_debug("crusde_register_param_string called for:%s", name);
 	ParamWrapper *p = new ParamWrapper();
+	p->setOptional(false);
 	SimulationCore::instance()->registerParam(p, name, category);
-	char** add = p->newString();
+	char** add =  p->newString();
+	return add;
+}
+
+extern "C" char** crusde_register_optional_param_string(const char* name, PluginCategory category, char* default_value)
+{
+crusde_debug("crusde_register_optional_param_string called for: %s\n", name);
+	ParamWrapper *p = new ParamWrapper();
+	p->setOptional(true);
+	SimulationCore::instance()->registerParam(p, name, category);
+	char** add =  p->newString(default_value);
 	return add;
 }
 
@@ -225,7 +304,9 @@ extern "C" const char* crusde_get_out_file()
 
 extern "C" void crusde_exit(int exitcode)
 {
-    SimulationCore::instance()->terminate();
+    if(SimulationCore::instance() != NULL)
+	    SimulationCore::instance()->terminate();
+
     exit(exitcode);
 }
 /*
